pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "owais-sid",email = "ahmedt2owais@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "openai (>=2.26.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "supabase (>=2.28.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
